+++ lib/src/code_generators/v3/swagger_enums_generator_v3.dart
@@ -1,10 +1,11 @@
 import 'dart:convert';
 import 'package:swagger_dart_code_generator/src/code_generators/swagger_enums_generator.dart';
+import 'package:swagger_dart_code_generator/src/code_generators/swagger_models_generator.dart';
 
 class SwaggerEnumsGeneratorV3 extends SwaggerEnumsGenerator {
   @override
   String generate(String dartCode, String fileName) {
-    final dynamic map = jsonDecode(dartCode);
+    final map = jsonDecode(dartCode) as Map<String, dynamic>;
 
     final components = map['components'] as Map<String, dynamic>?;
     final schemas = components == null
@@ -16,10 +17,18 @@ class SwaggerEnumsGeneratorV3 extends SwaggerEnumsGenerator {
         : components['responses'] as Map<String, dynamic>?;
 
     final requestBodies = components == null
-        ? null
-        : components['requestBodies'] as Map<String, dynamic>?;
+        ? <String, dynamic>{}
+        : components['requestBodies'] as Map<String, dynamic>? ?? {};
 
-    return generateFromMap(dartCode, fileName, schemas ?? {}, responses ?? {},
-        requestBodies ?? {});
+    requestBodies
+        .addAll(SwaggerModelsGenerator.getRequestBodiesFromRequests(map));
+
+    return generateFromMap(
+      dartCode,
+      fileName,
+      schemas ?? {},
+      responses ?? {},
+      requestBodies,
+    );
   }
 }