+++ lib/swagger_dart_code_generator.dart
@@ -4,6 +4,7 @@ import 'package:swagger_dart_code_generator/src/models/generator_options.dart';
 import 'package:swagger_dart_code_generator/src/swagger_code_generator.dart';
 import 'package:universal_io/io.dart';
 import 'package:dart_style/dart_style.dart';
+import 'package:path/path.dart' show join, normalize;
 
 ///Returns instance of SwaggerDartCodeGenerator
 SwaggerDartCodeGenerator swaggerCodeBuilder(BuilderOptions options) =>
@@ -18,26 +19,32 @@ const String _outputResponsesFileExtension = '.responses.swagger.dart';
 const String _indexFileName = 'client_index.dart';
 const String _mappingFileName = 'client_mapping.dart';
 
+String normal(String path) {
+  return AssetId('', path).path;
+}
+
 Map<String, List<String>> _generateExtensions(GeneratorOptions options) {
-  final filesList = Directory(options.inputFolder).listSync().where(
+  final filesList = Directory(normalize(options.inputFolder)).listSync().where(
       (FileSystemEntity file) =>
           _inputFileExtensions.any((ending) => file.path.endsWith(ending)));
 
   final result = <String, List<String>>{};
 
+  var out = normalize(options.outputFolder);
+
   filesList.forEach((FileSystemEntity element) {
     final name = getFileNameBase(element.path);
-    result[element.path] = <String>[
-      '${options.outputFolder}$name$_outputFileExtension',
-      '${options.outputFolder}$name$_outputEnumsFileExtension',
-      '${options.outputFolder}$name$_outputModelsFileExtension',
-      '${options.outputFolder}$name$_outputResponsesFileExtension',
+    result[normal(element.path)] = <String>[
+      join(out, '$name$_outputFileExtension'),
+      join(out, '$name$_outputEnumsFileExtension'),
+      join(out, '$name$_outputModelsFileExtension'),
+      join(out, '$name$_outputResponsesFileExtension'),
     ];
   });
 
   ///Register additional outputs in first input
-  result[filesList.first.path]!.add('${options.outputFolder}$_indexFileName');
-  result[filesList.first.path]!.add('${options.outputFolder}$_mappingFileName');
+  result[normal(filesList.first.path)]!.add(join(out, _indexFileName));
+  result[normal(filesList.first.path)]!.add(join(out, _mappingFileName));
 
   return result;
 }
@@ -103,8 +110,10 @@ class SwaggerDartCodeGenerator implements Builder {
 
     final dateToJson = codeGenerator.generateDateToJson(contents);
 
-    final copyAssetId = AssetId(buildStep.inputId.package,
-        '${options.outputFolder}$fileNameWithoutExtension$_outputFileExtension');
+    final copyAssetId = AssetId(
+        buildStep.inputId.package,
+        join(options.outputFolder,
+            '$fileNameWithoutExtension$_outputFileExtension'));
 
     if (!options.separateModels || !options.buildOnlyModels) {
       await buildStep.writeAsString(
@@ -124,8 +133,10 @@ class SwaggerDartCodeGenerator implements Builder {
       ///Write enums
       final formatterEnums = _tryFormatCode(enums);
 
-      final enumsAssetId = AssetId(buildStep.inputId.package,
-          '${options.outputFolder}$fileNameWithoutExtension$_outputEnumsFileExtension');
+      final enumsAssetId = AssetId(
+          buildStep.inputId.package,
+          join(options.outputFolder,
+              '$fileNameWithoutExtension$_outputEnumsFileExtension'));
 
       await buildStep.writeAsString(enumsAssetId, formatterEnums);
     }
@@ -140,8 +151,10 @@ class SwaggerDartCodeGenerator implements Builder {
         enums.isNotEmpty,
       ));
 
-      final enumsAssetId = AssetId(buildStep.inputId.package,
-          '${options.outputFolder}$fileNameWithoutExtension$_outputModelsFileExtension');
+      final enumsAssetId = AssetId(
+          buildStep.inputId.package,
+          join(options.outputFolder,
+              '$fileNameWithoutExtension$_outputModelsFileExtension'));
 
       await buildStep.writeAsString(enumsAssetId, formattedModels);
     }
@@ -200,7 +213,7 @@ $dateToJson
     final codeGenerator = SwaggerCodeGenerator();
 
     final indexAssetId =
-        AssetId(inputId.package, '${options.outputFolder}$_indexFileName');
+        AssetId(inputId.package, join(options.outputFolder, _indexFileName));
 
     final imports = codeGenerator.generateIndexes(swaggerCode, buildExtensions);
 
@@ -209,8 +222,8 @@ $dateToJson
     }
 
     if (options.withConverter && !options.buildOnlyModels) {
-      final mappingAssetId =
-          AssetId(inputId.package, '${options.outputFolder}$_mappingFileName');
+      final mappingAssetId = AssetId(
+          inputId.package, join(options.outputFolder, _mappingFileName));
 
       final mapping = codeGenerator.generateConverterMappings(
           swaggerCode, buildExtensions, hasModels);