+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 2.2.5
+
+* Fixed Issue ([#265](https://github.com/epam-cross-platform-lab/swagger-dart-code-generator/issues/265))
+* Fixed Issue ([#275](https://github.com/epam-cross-platform-lab/swagger-dart-code-generator/issues/275))
+* Returned possibility to use operationId for request name
+* Added generation of requestBodies integrated to requests, when it's complex model
+
+# 2.2.4
+
+* Fixes issues with allOf support
+* Fixed issue with basic types in responses models
+
 # 2.2.3
 
 * Fixed Issue ([#138](https://github.com/epam-cross-platform-lab/swagger-dart-code-generator/issues/138))